requirements.txt
azure-functions
coverage==5.5
mailchimp-marketing==3.0.36
pytest==6.2.2
requests==2.25.1
